+++ src/main/java/blackjack/domain/card/Card.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.Objects;
 
@@ -43,43 +43,5 @@ public class Card {
 
     public Shape getShape() {
         return shape;
-    }
-
-    // TODO 외부 class로 분리 할 것인지 논의, 접근 제어자 논의
-    public enum Value {
-        ACE(1),
-        TWO(2),
-        THREE(3),
-        FOUR(4),
-        FIVE(5),
-        SIX(6),
-        SEVEN(7),
-        EIGHT(8),
-        NINE(9),
-        TEN(10),
-        JACK(10),
-        QUEEN(10),
-        KING(10);
-
-        private final int minScore;
-
-        Value(int minScore) {
-            this.minScore = minScore;
-        }
-
-        public int getMinScore() {
-            return minScore;
-        }
-
-        public int getMaxScore() {
-            if (this == ACE) {
-                return minScore + 10;
-            }
-            return minScore;
-        }
-    }
-
-    public enum Shape {
-        SPADE, DIAMOND, HEART, CLOVER
     }
 }
