+++ src/main/java/blackjack/domain/card/Shape.java
@@ -1,0 +1,5 @@
+package blackjack.domain.card;
+
+public enum Shape {
+    SPADE, DIAMOND, HEART, CLOVER
+}
