+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,6 +1,6 @@
 package blackjack.domain.participant;
 
-import blackjack.domain.Card;
+import blackjack.domain.card.Card;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
