+++ src/test/java/blackjack/domain/CardTest.java
@@ -3,6 +3,9 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Shape;
+import blackjack.domain.card.Value;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -19,8 +22,8 @@ public class CardTest {
         @DisplayName("숫자 카드는 카드의 숫자 값을 점수로 가진다")
         @ParameterizedTest
         @CsvSource({"TWO, 2", "TEN, 10"})
-        void scoreTest_whenNumberCard(Card.Value value, int expected) {
-            Card card = new Card(value, Card.Shape.CLOVER);
+        void scoreTest_whenNumberCard(Value value, int expected) {
+            Card card = new Card(value, Shape.CLOVER);
 
             assertAll(
                     () -> assertThat(card.getMinScore()).isEqualTo(expected),
@@ -31,8 +34,8 @@ public class CardTest {
         @DisplayName("Jack, Queen, King 카드는 모두 10점을 가진다.")
         @ParameterizedTest
         @EnumSource(names = {"JACK", "QUEEN", "KING"})
-        void scoreTest_whenJQK_Card(Card.Value value) {
-            Card card = new Card(value, Card.Shape.HEART);
+        void scoreTest_whenJQK_Card(Value value) {
+            Card card = new Card(value, Shape.HEART);
 
             assertAll(
                     () -> assertThat(card.getMinScore()).isEqualTo(10),
@@ -43,7 +46,7 @@ public class CardTest {
         @DisplayName("Ace 카드는 1, 또는 11점을 가진다.")
         @Test
         void scoreTest_whenAceCard() {
-            Card card = new Card(Card.Value.ACE, Card.Shape.HEART);
+            Card card = new Card(Value.ACE, Shape.HEART);
 
             assertAll(
                     () -> assertThat(card.getMinScore()).isEqualTo(1),
