+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -3,10 +3,10 @@ package blackjack.domain.participant;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.Card;
-import blackjack.domain.Card.Shape;
-import blackjack.domain.Card.Value;
 import blackjack.domain.Deck;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Shape;
+import blackjack.domain.card.Value;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
